Conflicts:
	doc/demo/python/Test_FaciesIndicators.ipynb
	include/Basic/Utilities.hpp
	include/Basic/Vector.hpp
	include/Interfaces/AVariableTemplate.hpp
	include/License/LicenseKey.hpp
	include/License/LicenseUtility.hpp
	include/Stats/Selectivity.hpp
	python/modules/plot.py
	src/Basic/ASerializable.cpp
	src/Covariances/ACovAnisoList.cpp
	src/License/LicenseKey.cpp
	src/License/LicenseUtility.cpp
	src/OutputFormat/GridIfpEn.cpp
	src/Simulation/SimuSubstitutionParam.cpp
	src/all_sources.cmake
	swig/swig_exp.i
	swig/swig_inc.i
	tests/cpp/CMakeLists.txt
	tests/py/test_Matrix.py
	tests/py/test_Rotation.py

+++ src/OutputFormat/FileVTK.cpp
@@ -10,6 +10,7 @@
 /******************************************************************************/
 #include "geoslib_f.h"
 #include "geoslib_old_f.h"
+
 #include "OutputFormat/FileVTK.hpp"
 #include "OutputFormat/AOF.hpp"
 #include "Db/Db.hpp"
@@ -135,7 +136,7 @@ int FileVTK::writeInFile()
       if (! _db->isActive(iech)) continue;
       for (int idim = 0; idim < 3; idim++)
       {
-        double fact = 1.;
+        int fact = 1;
         if (idim == 0) fact = _factx;
         if (idim == 1) fact = _facty;
         if (idim == 2) fact = _factz;
